+++ .github/workflows/main.yml
@@ -16,10 +16,10 @@ jobs:
         dotnet-version: '7.0.x'
     
     - name: Publish .NET Core Project
-      run: dotnet publish src/SB12Calculator/SB12Calculator.csproj -c Release -o release --nologo
+      run: dotnet publish src/TaxTools/TaxTools.csproj -c Release -o release --nologo
         
-    - name: Change base-tag in index.html from / to SB12Calculator
-      run: sed -i 's/<base href="\/" \/>/<base href="\/SB12Calculator\/" \/>/g' release/wwwroot/index.html
+    - name: Change base-tag in index.html from / to TaxTools
+      run: sed -i 's/<base href="\/" \/>/<base href="\/TaxTools\/" \/>/g' release/wwwroot/index.html
         
     - name: copy index.html to 404.html
       run: cp release/wwwroot/index.html release/wwwroot/404.html
