+++ src/TaxTools.Core/Calculator.cs
@@ -1,4 +1,4 @@
-﻿namespace SB12Calculator.Core
+﻿namespace TaxTools.Core
 {
     public static class Calculator
     {
