+++ src/TaxTools.Core/TaxTools.Core.csproj
