+++ src/TaxTools/_Imports.razor
@@ -6,5 +6,5 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
-@using SB12Calculator
-@using SB12Calculator.Shared
+@using TaxTools
+@using TaxTools.Shared
